[0.6.6] feat(Web) Added debut rating album

+++ Musicx.Infrastructure/API/Persistence/Schema/Albums/CreateAlbumRatingStatTable.sql
@@ -7,6 +7,6 @@ drop table if exists public.album_rating_stats cascade;
 create table public.album_rating_stats (
     album_rating_stats_album_id             bigint primary key references public.albums(album_id) on delete cascade,
     album_rating_stats_count                integer not null default 0,
-    album_rating_stats_sum                  integer not null default 0,
+    album_rating_stats_sum                  bigint not null default 0,
     album_rating_stats_avg                  numeric(5,2) not null default 0
 );
